Remove a classe Main e altera o tipo do ID de Long para int na classe Expense. Atualiza ExpenseService para carregar despesas do repositório, renomeia métodos e adiciona funcionalidades para resumo e exclusão de despesas.

+++ expense-tracker/src/main/java/com/nullifidianz/model/Expense.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class Expense {
 
-    private Long id;
+    private int id;
     private String description;
     private Double amount;
     private LocalDate date;
@@ -13,18 +13,18 @@ public class Expense {
 
     }
 
-    public Expense(Long id, String description, Double amount, LocalDate date) {
+    public Expense(int id, String description, Double amount, LocalDate date) {
         this.id = id;
         this.description = description;
         this.amount = amount;
         this.date = date;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
